feat(pastebin): improve result box
