+++ apps/ui/src/app/dashboard/pages/pastes/pastes.component.ts
@@ -7,7 +7,6 @@ import {
     MenuItem,
     MenuItemCommandEvent,
 } from 'primeng/api';
-import dayjs from 'dayjs';
 import { Menu } from 'primeng/menu';
 import { ToastService } from '../../../shared/services/toast/toast.service';
 import { Page } from '../../../shared/interfaces/http';
@@ -16,6 +15,7 @@ import {
     Paste,
     PastebinService,
 } from '../../../pastebin/services/pastebin.service';
+import { DateUtils } from '../../../shared/utils/date';
 
 @Component({
     selector: 'app-pastes',
@@ -101,19 +101,6 @@ export class PastesComponent {
         });
     }
 
-    pasteExpiresIn(expiresAt: string, format: 'humanized' | 'long') {
-        if (format === 'humanized') {
-            const expiration = dayjs(expiresAt);
-            if (expiration < dayjs()) {
-                return 'Expired';
-            } else {
-                return expiration.fromNow();
-            }
-        } else {
-            return dayjs(expiresAt).format('MMMM D, YYYY h:mm A');
-        }
-    }
-
     delete(id: number) {
         this.pastebinService.delete(id).subscribe({
             next: () => {
@@ -166,4 +153,6 @@ export class PastesComponent {
         });
         menu.toggle(event);
     }
+
+    protected readonly DateUtils = DateUtils;
 }
