+++ apps/ui/src/app/pastebin/pages/index/pastebin.component.html
@@ -87,18 +87,30 @@
 
     <div
         *ngIf="paste"
+        [@inOutAnimation]
         class="flex flex-col lg:flex-row bg-red-700 mt-5 border-2 border-red-300 rounded p-3"
     >
         <qrcode [qrdata]="URL" [width]="120"></qrcode>
         <div class="flex-1 flex flex-col justify-center px-3 text-xl">
-            <span class="flex align-center gap-2 font-bold mt-2 lg:mt-0">
+            <span class="flex align-center font-bold mt-2 lg:mt-0">
                 <i
                     pTooltip="Paste URL"
                     tooltipPosition="left"
                     class="ti ti-link text-red-300"
                     style="font-size: 28px"
                 ></i>
+                <span class="text-red-300 ms-1 mr-3">URL</span>
                 <a id="paste-url" class="text-red-100" [href]="URL">{{ URL | slice: 8 }}</a>
+            </span>
+            <span class="flex align-center font-bold mt-2 lg:mt-0">
+                <i
+                    [pTooltip]="DateUtils.expiresIn(paste!.expiresAt, 'long')"
+                    tooltipPosition="left"
+                    class="ti ti-calendar text-red-300"
+                    style="font-size: 28px"
+                ></i>
+                <span class="text-red-300 ms-1 mr-3">Expires</span>
+                <span id="paste-expiration" class="text-red-100">{{ DateUtils.expiresIn(paste!.expiresAt, 'humanized') }}</span>
             </span>
         </div>
         <p-button
